StarMap v5
this is my 5th commit to git hub.
in this commit i attempted to add the text of the co-ordinates to the
map. the code i commented out due to it not working correctly. it is now
after the time limit of the exam so this is my final commit and my
submission.